Create LinearSearch.py
Linear Search is a type of searching algorithm. Implemented both ordered linear search for searching in an unordered list. Time Complexity = O(n) 
For ordered Linear search, we first sort the list then search for an element, if the current element being iterated by the for loop is greater than the search value, then we know that what we are to search is not in the list and return false. Time Complexity = O(n) but more efficient than unordered linear search. 

Applications: When we need to determine the position of a search element without sorting the list. 

Useful For small data sets, otherwise for larger data sets to search, we can do more better than linear time complexity. We gat to think of another better way, Binary search that takes O(log n) time.